WIP: added top app bar route
