+++ test/screenshot/index.js
@@ -1,12 +1,14 @@
+import path from 'path';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import {BrowserRouter, Route} from 'react-router-dom';
-
+import {HashRouter, Route} from 'react-router-dom';
 import Home from './Home';
-import TextField from './text-field/index';
+import topAppBarVariants from './top-app-bar/variants';
+
+import './index.scss';
 
 ReactDOM.render((
-  <BrowserRouter>
+  <HashRouter>
     <div>
       <Route exact path='/' component={Home}/>
       {COMPONENTS.map((componentPath) => {
@@ -15,6 +17,13 @@ ReactDOM.render((
           <Route exact key={componentPath} path={`/${componentPath}`} component={Component}/>
         );
       })}
+      {topAppBarVariants.map((variant) => {
+        const path = `top-app-bar/${variant}`;
+        const Component = require(path.resolve(__dirname, `${path}.js`)).default;
+        return (
+          <Route exact key={path} path={`/${path}`} component={Component}/>
+        );
+      })}
     </div>
-  </BrowserRouter>
+  </HashRouter>
 ), document.getElementById('mcdr-screenshot-test-app'));
