+++ test/screenshot/index.scss
@@ -1,0 +1,5 @@
+body {
+  font-family: 'Roboto', Arial, sans-serif;
+  margin: 0;
+  padding: 0;
+}
