+++ test/screenshot/top-app-bar/prominent.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import TopAppBar from '../../../packages/top-app-bar';
 
 import MaterialIcon from '../../../packages/material-icon';
