+++ test/screenshot/top-app-bar/variants.js
@@ -1,0 +1,9 @@
+export default [
+  'fixed',
+  'prominent',
+  'short',
+  'shortCollapsed',
+  'standard',
+  'standardNoActionItems',
+  'standardWithNavigationIconElement',
+];
